requirements.txt
ripe.atlas.cousteau
numpy
